Split docker image in generic base and default
The current image makes some assumptions that are hard to undo
downstream, namely the EXPOSE and VOLUME lines (and to some
degree: The ENTRYPOINT). By splitting the file into a binary base
image and a default configuration derived by it, downstream can
use

FROM difi/oxalis-base

in the future to customize the distribution further, while current
users of the difi/oxalis image will see exactly the same behavior as
today.